event_manager/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements-dev.in
#
asgiref==3.7.2
    # via
    #   -c requirements.txt
    #   django
django==4.2.2
    # via
    #   -c requirements.txt
    #   django-debug-toolbar
django-debug-toolbar==4.1.0
    # via -r requirements-dev.in
factory-boy==3.2.1
    # via -r requirements-dev.in
faker==18.10.1
    # via factory-boy
python-dateutil==2.8.2
    # via faker
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via
    #   -c requirements.txt
    #   django
    #   django-debug-toolbar
tzdata==2023.3
    # via
    #   -c requirements.txt
    #   django
